pyproject.toml
[project]
name = "gpt-sovits"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "av>=11",
    "chardet>=7.4.3",
    "cn2an>=0.5.23",
    "ctranslate2>=4.0,<5",
    "fast-langdetect>=0.3.1",
    "fastapi[standard]>=0.115.2",
    "ffmpeg-python>=0.2.0",
    "funasr==1.0.27",
    "g2p-en>=2.1.0",
    "g2pk2>=0.0.3",
    "gradio<5",
    "jieba>=0.42.1",
    "jieba-fast>=0.53",
    "ko-pron>=1.3",
    "librosa==0.10.2",
    "modelscope>=1.35.4",
    "numba>=0.65.0",
    "numpy<2.0",
    "onnxruntime>=1.24.3 ; platform_machine == 'aarch64' or platform_machine == 'arm64'",
    "onnxruntime-gpu>=1.24.3 ; platform_machine == 'AMD64' or platform_machine == 'x86_64'",
    "opencc>=1.2.0",
    "peft<0.18.0",
    "psutil>=7.2.2",
    "pydantic<=2.10.6",
    "pyopenjtalk>=0.4.1",
    "pypinyin>=0.55.0",
    "python-dotenv>=1.2.2",
    "python-mecab-ko>=1.3.7 ; sys_platform != 'win32'",
    "pytorch-lightning>=2.4",
    "pyyaml>=6.0.3",
    "rotary-embedding-torch>=0.8.9",
    "scipy>=1.15.3",
    "sentencepiece>=0.2.1",
    "split-lang>=2.1.1",
    "tensorboard>=2.20.0",
    "tojyutping>=3.2.0",
    "torchaudio>=2.11.0",
    "torchmetrics<=1.5",
    "tqdm>=4.67.3",
    "transformers>=4.43,<=4.50",
    "wordsegment>=1.3.1",
    "x-transformers>=2.18.8",
]
